Deleted closer_utils.py, moved everything to DataServiceUtils.py as agreed in the review
